\name{BetaBinomial}
\alias{BetaBinomial}

\title{
  Shiny App to Illustrate Posterior Calculations for a Beta-Binomial Model
}
\description{
  Interactively choose dataset, second-stage  prior parameters, and number of simulations and app illustrates posterior calculations
}
\usage{
  BetaBinomial()
}
\author{
  Jim Albert
}
