\name{BetaBinomialJAGS}
\alias{BetaBinomialJAGS}

\title{
  Shiny App to Illustrate Posterior Calculations for a Beta-Binomial Model using JAGS
}
\description{
  Interactively choose dataset, second-stage  prior parameters, and number of simulations and app illustrates posterior calculations
}
\usage{
  BetaBinomialJAGS()
}
\author{
  Jim Albert
}
