\name{Discrete_TwoP}
\alias{Discrete_TwoP}

\title{
  Shiny App to Illustrate Posterior Calculations for Two Proportions with a Discrete Prior
}
\description{
  Interactively choose dataset and parameters of discrete prior,  and app illustrates posterior calculations
}
\usage{
  Discrete_TwoP()
}
\author{
  Jim Albert
}
