\name{HowardPrior}
\alias{HowardPrior}

\title{
  Shiny App to Visualize Comparison of Two Proportions with Howard's Prior
}
\description{
  Interactively choose prior parameter and number of successes and failures in two samples and app displays posterior of two proportions
}
\usage{
  HowardPrior()
}
\author{
  Jim Albert
}
