\name{NormalNormal}
\alias{NormalNormal}

\title{
  Shiny App to Illustration Posterior Calculations for a Normal-Normal Multilevel Model
}
\description{
  Interactively choose dataset, second-stage  prior parameters, and number of simulations and app illustrates posterior calculations
}
\usage{
  NormalNormal()
}
\author{
  Jim Albert
}
