\name{TwoProportions}
\alias{TwoProportions}

\title{
  Shiny App to Visualize Comparison of Two Proportions
}
\description{
  Interactively choose  number of successes and failures in two samples and app shows the posterior of the difference and sum of the corresponding logits
}
\usage{
  TwoProportions()
}
\author{
  Jim Albert
}
